Adicionando permissões nas tasks

+++ src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from './../auth/user.entity';
 import { TaskRepository } from './task.repository';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -13,12 +14,18 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) { }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto);
+  async getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User,
+  ): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User,
+  ): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
       throw new NotFoundException(`Task with id "${id}" not found`);
@@ -27,12 +34,18 @@ export class TasksService {
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-    async deleteTask(id: number): Promise<void> {
-      const result = await this.taskRepository.delete(id);
+    async deleteTask(
+      id: number,
+      user: User,
+    ): Promise<void> {
+      const result = await this.taskRepository.delete({ id, userId: user.id });
 
       if (result.affected === 0) {
         throw new NotFoundException(`Task with id "${id}" not found`);
@@ -46,8 +59,12 @@ export class TasksService {
   //   await this.taskRepository.delete(id);
   // }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
